requirements.txt
pycryptodome
pywin32
prettytable
